New mechanism to ensure actual changes
We have changed the whole logic in order to ensure that a change is
actually a change. This happens because when you use the dot notation
to update a path, and there are arrays in the path, mongoose does not
call $set with the new value, so we have to mark the change as unchecked
and check it after the execution of the proxy middleweares finish.

In order to achieve this we must send an event that will execute an
immediate with the checker function and this checker function will check
if the unchecked changes are actual changes or not, or if the changes
are of complex objects, then we will check which paths are actually
changed, and we will insert as changes the paths that have actaully
changed.

This way we can ensure that the changes array contains actual changes.
and not only assignations as before.