Don't check for granular changes
In the previous version we implemented the ability to convert a single
change to a bunch of changes only for the deepest actual changes.
This that could sound great it is not so great because when you are
setting a new path with a big object value, the actual changes may be a
lot, generating an ammount of changes that should be kept as a single
actual change.

This was done in order to detect only actual changes, but it is not
efficient enough.