Update how old_value is calculated [README]
In most cases we are going to take just the current value of the given
path, but, if we come from a mutable array method call we create a copy
of the array because otherwise we are going to store a reference to the
array, and when the mutable array method is called, the array will
mutate and the old_value, as it is a reference, will be the array with
the applied changed, so we could not compare the old and the new values.

We don't want to make a copy in every situation because we don't want to
create more memmory for most cases.