Create a separate markModified proxy
Before these changes, the proxy that controlled both real changes and
calls to the markModified function was the same. That doesn't make much
sense, as all the calls for markModified won't enter the if as they
only have 1 argument (the only have path, not value).

This commit creates a custom markModified proxy that only adds the path
for a future changes check, as a markModified call is supposed to be.

It is important to know that the markModified function must be called
with the path to be modified before it is actually modified, otherwise
the changes will not be detected.

This commit also adds a bunch of tests for this particular use case,
attacking a Mixed type that mongoose is unable to track the changes
and the markModified should be called manually in order to detect
the changes.